chore: make jest results visible in Storybook
