+++ .github/workflows/chromatic.yml
@@ -40,6 +40,9 @@ jobs:
         if: steps.npm-cache.outputs.cache-hit != 'true'
         run: npm ci
 
+      - name: Generate the Jest output
+        run: npm run test:generate-output
+
         #👇 Adds Chromatic as a step in the workflow
       - uses: chromaui/action@v1
         # Options required for Chromatic's GitHub Action
