+++ .storybook/main.ts
@@ -9,6 +9,7 @@ module.exports = {
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
     "@storybook/addon-a11y",
+    "@storybook/addon-jest",
     "./ouia-addon/preset",
   ],
   features: {
