+++ .storybook/preview.ts
@@ -8,12 +8,14 @@ import "@patternfly/patternfly/utilities/Sizing/sizing.css";
 import "@patternfly/patternfly/utilities/Spacing/spacing.css";
 import "@patternfly/patternfly/utilities/Text/text.css";
 import "./ouia-helper.css";
-import { inspect } from "@xstate/inspect";
-import React, { useEffect } from "react";
-import { BrowserRouter as Router } from "react-router-dom";
-import { AppServicesLoading, I18nProvider } from "../src";
+import { Parameters, ReactFramework } from "@storybook/react/types-6-0";
+import { DecoratorFunction } from "@storybook/csf";
+import { withTests } from "@storybook/addon-jest";
 
-export const parameters = {
+import results from "../.jest-test-results.json";
+import { withMas } from "./withMas";
+
+export const parameters: Parameters = {
   options: {},
   previewTabs: { "storybook/docs/panel": { index: -1 } },
   ouia: "false",
@@ -131,58 +133,10 @@ export const globalTypes = {
   },
 };
 
-export const decorators = [
-  (Story, { globals }) => {
-    useEffect(() => {
-      document.body.classList.toggle("show-ouia", JSON.parse(globals.ouia));
-    }, [globals.ouia]);
-    useEffect(() => {
-      let inspector;
-
-      if (JSON.parse(globals.xstate) === true) {
-        inspector = inspect({
-          // options
-          url: "https://stately.ai/viz?inspect", // (default)
-          iframe: false, // open in new window
-        });
-      }
-      return () => {
-        if (inspector) {
-          inspector.disconnect();
-          inspector = undefined;
-        }
-      };
-    }, [globals.xstate]);
-    return (
-      <Router>
-        <I18nProvider
-          lng={globals.locale}
-          resources={{
-            en: {
-              common: () => import("../locales/en/common.json"),
-              "create-kafka-instance": () =>
-                import("../locales/en/create-kafka-instance.json"),
-              kafka: () => import("../locales/en/kafka.json"),
-              metrics: () => import("../locales/en/metrics.json"),
-              overview: () => import("../locales/en/overview.json"),
-              datascienceoverview: () =>
-                import("../locales/en/datascienceoverview.json"),
-              apimgmtoverview: () =>
-                import("../locales/en/apimgmtoverview.json"),
-              "manage-kafka-permissions": () =>
-                import("../locales/en/manage-kafka-permissions.json"),
-            },
-            it: {
-              common: () => Promise.resolve({ delete: "Elimina" }),
-            },
-          }}
-          debug={true}
-        >
-          <React.Suspense fallback={<AppServicesLoading />}>
-            <Story />
-          </React.Suspense>
-        </I18nProvider>
-      </Router>
-    );
-  },
+export const decorators: DecoratorFunction<ReactFramework>[] = [
+  withTests({
+    results,
+    filesExt: ".test.tsx",
+  }),
+  withMas,
 ];
